Shim for geometric_dimension, break circular dependencies
Move a couple functions from utilities.py into calculus.py so that we
avoid circular dependencies. This shouldn't affect downstream code
because they should only be importing those functions from the top-level
`icepack` namespace.
