+++ src/icepack/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2017-2023 by Daniel Shapero <shapero@uw.edu>
+# Copyright (C) 2017-2026 by Daniel Shapero <shapero@uw.edu>
 #
 # This file is part of icepack.
 #
@@ -12,7 +12,8 @@
 
 from icepack.norms import norm
 from icepack.interpolate import interpolate
-from icepack.utilities import depth_average, lift3d, compute_surface, vertical_velocity
+from icepack.utilities import compute_surface
+from icepack.calculus import vertical_velocity, depth_average, lift3d
 from icepack.models.viscosity import rate_factor
 import icepack.meshing
 import icepack.datasets
